Separates some of the logic for binding custom shortcuts and standard collection rules; also simplifies the syntax used when binding standard collection rules (to be able to skip over custom shortcuts)
